:recycle: Extract wasm-functions for fills and strokes out of main.rs
