+++ render-wasm/src/wasm.rs
@@ -1,2 +1,4 @@
+pub mod fills;
 pub mod fonts;
+pub mod strokes;
 pub mod text;
